original-backend/requirements.txt
APScheduler==3.9.1
arabic-reshaper==2.1.3
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.2
django-bulk-update==2.2.0
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
executing==0.8.2
fonttools==4.29.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-resources==5.4.0
ipython==8.0.1
ipython-genutils==0.2.0
jedi==0.18.1
jsonschema==4.4.0
jupyter-core==4.9.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nbformat==5.1.3
numpy==1.22.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
plotly==5.5.0
prompt-toolkit==3.0.27
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.11.2
pyparsing==3.0.7
PyPDF2==1.26.0
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.6
requests==2.27.1
retrying==1.3.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tenacity==8.0.1
toml==0.10.2
tomli==2.0.0
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
zipp==3.7.0
